server/requirements/base.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers>=4.6,<5.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
openpyxl>=3.1,<4.0
PyJWT>=2.8,<3.0
sqlparse==0.5.5
